fix: оптимизация и исправление замечаний

+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/db/UserDbStorage.java
@@ -2,15 +2,17 @@ package ru.yandex.practicum.filmorate.storage.impl.db;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.impl.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 import ru.yandex.practicum.filmorate.storage.impl.db.converter.UserConverter;
 
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @Component
 @Primary
@@ -20,25 +22,16 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public List<User> findAll() {
-        List<User> users = new ArrayList<>();
-        UserConverter userConverter = new UserConverter();
-        SqlRowSet userRows = jdbcTemplate.queryForRowSet("select * from users order by user_id");
-        while (userRows.next()) {
-            User user = userConverter.apply(userRows);
-            setFriendsForInstance(user);
-            users.add(user);
-        }
-        return users;
+        return jdbcTemplate.query("select * from users order by user_id", (rs, rn) -> UserConverter.fromResultSet(rs));
     }
 
     @Override
     public Optional<User> findById(int id) {
-        SqlRowSet userRow = jdbcTemplate.queryForRowSet("select * from users where user_id = ?", id);
-        if (userRow.next()) {
-            User user = new UserConverter().apply(userRow);
-            setFriendsForInstance(user);
-            return Optional.of(user);
-        } else {
+        try {
+            return jdbcTemplate.queryForObject("select * from users where user_id = ?",
+                    (rs, rn) -> Optional.of(UserConverter.fromResultSet(rs)),
+                    id);
+        } catch (EmptyResultDataAccessException ex) {
             return Optional.empty();
         }
     }
@@ -49,7 +42,14 @@ public class UserDbStorage implements UserStorage {
                 .withTableName("users")
                 .usingGeneratedKeyColumns("user_id");
 
-        int id = simpleJdbcInsert.executeAndReturnKey(user.toMap()).intValue();
+        Map<String, Object> values = Map.of(
+                "email", user.getEmail(),
+                "login", user.getLogin(),
+                "name", Optional.ofNullable(user.getName()).orElse(""),
+                "birthday", user.getBirthday()
+        );
+
+        int id = simpleJdbcInsert.executeAndReturnKey(values).intValue();
         user.setId(id);
 
         return user;
@@ -73,29 +73,44 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public void addFriend(User user, User friend) {
-        user.addFriend(friend);
         jdbcTemplate.update("insert into friends values(?, ?, 2)", user.getId(), friend.getId());
     }
 
     @Override
     public void deleteFriend(User user, User friend) {
-        user.deleteFriend(friend);
-        jdbcTemplate.update("delete from friends " +
-                        " where first_user_id = ? " +
-                        "   and second_user_id = ? ",
-                user.getId(), friend.getId());
+        jdbcTemplate.update("delete from friends where first_user_id = ? and second_user_id = ?",
+                user.getId(),
+                friend.getId());
     }
 
-    private void setFriendsForInstance(User user) {
-        Set<User> friends = new HashSet<>();
-        SqlRowSet friendRow = jdbcTemplate.queryForRowSet("select f.second_user_id " +
+    @Override
+    public List<User> findFriends(int id) {
+        return jdbcTemplate.query("select " +
+                        " u.* " +
                         " from friends f " +
-                        " where f.first_user_id = ? and f.frstat_frstat_id = 2",
-                user.getId());
-        while (friendRow.next()) {
-            Optional<User> userOpt = findById(friendRow.getInt("second_user_id"));
-            userOpt.ifPresent(friends::add);
-        }
-        user.setFriends(friends);
+                        " join users u " +
+                        "   on u.user_id = f.second_user_id " +
+                        " where f.first_user_id = ? " +
+                        "   and f.frstat_frstat_id = 2 " +
+                        " order by u.user_id",
+                (rs, rn) -> UserConverter.fromResultSet(rs),
+                id);
+    }
+
+    @Override
+    public List<User> findCommonFriends(int id, int otherId) {
+        return jdbcTemplate.query("select " +
+                        " u.* " +
+                        " from (" +
+                        "     select f.second_user_id from friends f where f.first_user_id = ? and f.frstat_frstat_id = 2 " +
+                        "     intersect " +
+                        "     select f.second_user_id from friends f where f.first_user_id = ? and f.frstat_frstat_id = 2 " +
+                        " ) fr " +
+                        " join users u " +
+                        "   on u.user_id = fr.second_user_id " +
+                        " order by u.user_id",
+                (rs, rn) -> UserConverter.fromResultSet(rs),
+                id,
+                otherId);
     }
 }
